requirements.txt
django==2.1
mysqlclient==1.3.13
django-fullurl==0.5
django-uuslug==1.1.8
